Fixed Highlight bug w/ NavBar
Fixed bug where "Home" was always highlighted, no matter which tab the user was on. Interestingly, the problem was the lines ".menuitems a:hover, .menuitems a:active ". changing a:active to a.active fixes these changes, because a:active applies when the link is being clicked, not when it is the current page. The current page automatically gets updated with an "active" class, so calling a.active in css will apply the proper changes.