Fixed Image overlap on HomePage
Fixed issue where the image of the Tesla would cover the navigation bar. The fix was to make the image a background image, that way everything is placed on top of it. While doing this, I also edited body's margins to be 0, in order to make the image fit to the entire page.