Created Login Page
Created login page with email and password input. Put button in the profile page to go access the login page.

Note: for pages that import ChakraUI components, the font in the navbar switches to the chakra ui font. Non-chakraUI pages still have the default Times New Roman font in the navbar.